+++ src/js/reducers/tokensReducer.js
@@ -147,6 +147,7 @@ const tokens = (state = initState, action) => {
       const {hash, symbol} = action.payload
       var tokens = { ...state.tokens }
       tokens[symbol].approveTx = hash
+      console.log(tokens)
       return Object.assign({}, state, { tokens: tokens }) 
     }
     case 'EXCHANGE.REMOVE_APPROVE_TX':{