+++ src/js/services/ethereum/nodeProviders/BaseProvider.js
@@ -33,12 +33,8 @@ export default class BaseProvider {
 
     getLatestBlock() {
         return new Promise((resolve, reject) => {
-            this.rpc.eth.getBlock("latest", false).then((block) => {
-                if (block != null) {
+            this.rpc.eth.getBlockNumber().then((block) => {
                     resolve(block)
-                } else {
-                    reject(new Error("Cannot get latest block from: " + this.rpcUrl))
-                }
             }).catch((err) => {
                 reject(err)
             })
@@ -188,10 +184,14 @@ export default class BaseProvider {
 
     exchangeData(sourceToken, sourceAmount, destToken, destAddress,
         maxDestAmount, minConversionRate, walletId) {
+
+        if (!this.rpc.utils.isAddress(walletId)){
+            walletId = "0x" + Array(41).join("0")
+        }  
         var data = this.networkContract.methods.trade(
             sourceToken, sourceAmount, destToken, destAddress,
             maxDestAmount, minConversionRate, walletId).encodeABI()
-
+                 
         return new Promise((resolve, reject) => {
             resolve(data)
         })
@@ -272,6 +272,20 @@ export default class BaseProvider {
                 })
         })
     }
+
+    // getRateTest(source, dest, quantity) {
+    //     return new Promise((resolve, reject) => {
+    //         this.networkContract.methods.getExpectedRate(source, dest, quantity).call()
+    //             .then((result) => {
+    //                 if (result != null) {
+    //                    // resolve(result)
+    //                 }
+    //             })
+    //             .catch((err) => {
+    //                 reject(err)
+    //             })
+    //     })
+    // }
 
     checkKyberEnable() {
         return new Promise((resolve, reject) => {
@@ -432,15 +446,18 @@ export default class BaseProvider {
 
         })
     }
-
+    
+    
     exactTradeData(data) {
         return new Promise((resolve, reject) => {
             try {
                 //get trade abi from 
                 var tradeAbi = this.getAbiByName("trade", constants.KYBER_NETWORK)
+              //  console.log(tradeAbi)
                 abiDecoder.addABI(tradeAbi)
+              //  console.log(abiDecoder)
                 var decoded = abiDecoder.decodeMethod(data);
-                //  console.log(decoded)
+            //      console.log(decoded)
                 resolve(decoded.params)
             } catch (e) {
                 reject(e)
@@ -506,8 +523,10 @@ export default class BaseProvider {
         return new Promise((resolve, reject) => {
             this.rpc.eth.getGasPrice()
                 .then(result => {
+                    
+                    console.log(result)
                     var gasPrice = parseInt(result, 10)
-                    if (gasPrice > 2000000000) {
+                    if (gasPrice > 20000000000) {
                         resolve({
                             low: "20",
                             default: "20",
@@ -516,12 +535,22 @@ export default class BaseProvider {
                         })
                     } else {
                         gasPrice = gasPrice / 1000000000
-                        resolve({
-                            low: (gasPrice * 0.7).toString(),
-                            default: gasPrice.toString(),
-                            standard: gasPrice.toString(),
-                            fast: (gasPrice * 1.7).toString()
-                        })
+                        if (gasPrice < 1){
+                            resolve({
+                                low: 1,
+                                default: 1,
+                                standard: 1,
+                                fast: 1
+                            })
+                        }else{
+                            resolve({
+                                low: gasPrice.toString(),
+                                default: gasPrice.toString(),
+                                standard: gasPrice.toString(),
+                                fast: (gasPrice * 1.7).toString()
+                            })
+                        }
+                        
                     }
                 }).catch((err) => {
                     reject(err)
@@ -565,22 +594,32 @@ export default class BaseProvider {
         })
     }
 
-    wrapperGetConversionRate(reserve, input, blockno) {
-        var data = this.networkContract.methods.getExpectedRate(input.source, input.dest, input.srcAmount).encodeABI()
+    getRateAtSpecificBlock(source, dest, srcAmount, blockno) {
+        var data = this.networkContract.methods.getExpectedRate(source, dest, srcAmount).encodeABI()
+        
         return new Promise((resolve, reject) => {
             this.rpc.eth.call({
                 to: BLOCKCHAIN_INFO.network,
                 data: data
             }, blockno)
                 .then(result => {
-                    var rates = this.rpc.eth.abi.decodeParameters([{
-                        type: 'uint256',
-                        name: 'expectedPrice'
-                    }, {
-                        type: 'uint256',
-                        name: 'slippagePrice'
-                    }], result)
-                    resolve(rates)
+                    console.log("rate: " + result)
+                    try{
+                        var rates = this.rpc.eth.abi.decodeParameters([{
+                            type: 'uint256',
+                            name: 'expectedPrice'
+                        }, {
+                            type: 'uint256',
+                            name: 'slippagePrice'
+                        }], result)
+                        console.log(rates)
+                        resolve(rates)
+                    }catch(e){
+                        resolve({
+                            expectedPrice: "0",
+                            slippagePrice: "0"
+                        })
+                    }
                 }).catch((err) => {
                     reject(err)
                 })